A count is an integer and an ordering is a comparison
Three places where the right answer rested on a rounding, all of them written
in the last two changes, none of them producing a wrong answer on any mode
this tree has seen.

The refresh estimator counted refreshes in a double and rounded with
floor(x + 0.5), which has the classic flaw at 0.49999999999999994 -- harmless
here because the quarter-refresh tolerance rejects it first, and llround into
a uint64 is free. The mode policy decided whether a mode drops frames with
refresh.hz() / fps.hz() < 1.0, and broke a tie by comparing two doubles where
two rationals a hair apart can land on the same one. Both are cross
multiplications now, the way Rational::operator== always was.

The first attempt at the second of those was itself wrong: numerator <
2 * denominator, where the denominator is already a product of two 32-bit
numbers and doubling it can leave 64 bits. Comparing the two rationals
directly has each side a product of two 32-bit numbers and nothing to
overflow.

What stays floating point is what should be: the gap is a measurement, the
seconds accumulate to a relative error of 2e-11 over an hour, and
refreshes_per_frame and seconds_between_slips are reported rather than
decided on.

362 tests pass at Release/AVX2.

+++ src/win/mediaperch/main.cpp
@@ -1867,8 +1867,9 @@ int show(const MpSinkVtbl& sink_vtbl, const mp::win::ModuleRegistry& registry,
     // estimator this replaced read a sixteen-and-two-thirds millisecond refresh
     // as sixteen and a half. The span is how many refreshes it averaged over,
     // which is what says how much of that third decimal to believe.
-    std::printf("refresh    %.3f ms measured, over %.0f refreshes\n",
-                turns.refresh_seconds * 1000.0, turns.refresh_span);
+    std::printf("refresh    %.3f ms measured, over %llu refreshes\n",
+                turns.refresh_seconds * 1000.0,
+                static_cast<unsigned long long>(turns.refresh_span));
     std::printf("turns      %llu, %llu of them with no clock to read\n",
                 static_cast<unsigned long long>(turns.turns),
                 static_cast<unsigned long long>(turns.without_clock));
